created files to separate pages
